Option to filter based on Group
With this change you can set a filter to do something based on the original group of the item.  For example if your source m3u labels some channels as 'Sport' you can set a filter to change the group to 'Action' etc.  It may be possible to use a wildcard * as well which would be cool but I haven't tested this yet.